saas-web-audit/backend/requirements.txt

fastapi==0.115.5
uvicorn[standard]==0.32.0
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
loguru==0.7.2
reportlab==4.2.5
jinja2==3.1.4
SQLAlchemy==2.0.34
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
celery==5.4.0
redis==5.0.1
python-multipart==0.0.9
